Enable users to connect to a remote lnd instance.

+++ public/electron.js
@@ -6,6 +6,7 @@ const url = require('url');
 const isDev = require('electron-is-dev');
 const log = require('electron-log');
 const { startLndProcess, startBtcdProcess } = require('./lnd-child-process');
+const { parseCliArg } = require('../public/helper');
 const grcpClient = require('./grpc-client');
 const {
   PREFIX_NAME,
@@ -37,6 +38,7 @@ const btcdSettingsDir = path.join(isDev ? 'data' : userDataPath, 'btcd');
 const lndArgs = process.argv.filter(a =>
   /(^--bitcoin)|(^--btcd)|(^--neutrino)/.test(a)
 );
+let lndIP = parseCliArg('lndip');
 
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
@@ -125,12 +127,21 @@ function createWindow() {
   grcpClient.init({
     ipcMain,
     lndSettingsDir,
+    lndIP,
     lndPort: LND_PORT,
     network: isDev ? 'simnet' : NETWORK,
   });
 }
 
 ipcMain.on('lnd-restart-process', async event => {
+  if (lndIP) {
+    event.sender.send(
+      'lnd-restart-error',
+      'Unable to reset password of remote lnd node.'
+    );
+    return;
+  }
+
   lndProcess && lndProcess.kill('SIGINT');
   let restartError;
   try {
@@ -199,7 +210,7 @@ app.on('ready', () => {
   initAutoUpdate();
   createWindow();
   initApplicationMenu();
-  startLnd();
+  if (!lndIP) startLnd();
 });
 
 // Quit when all windows are closed.
